backend/requirements.txt
Django>=2.0,<=3.0
psycopg2>=2.7,<=3.0 
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
django-cors-headers==3.5.0
PyYAML==5.3.1
uritemplate==3.0.1
